Added Rudimentary Collision
Basic sphere-based collision has been added for collision with obstacles, such as the lamps. This stops the car immediately.
A simple bounding-box has been added to determine if the car is on the road, but no punishment has been added for driving off the road.
Both of these features print to the console when they occur.

Also changed some variables into constants and put them in their own section so they may be changed easier for testing purposes. (Things like MAX_SPEED and CAR_COLLIDER_RADIUS for example)